fix: store structure
